% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mp.R
\name{fit_mp}
\alias{fit_mp}
\title{Fit the Marchenko-Pastur distribution}
\usage{
fit_mp(expr, sample = FALSE, cor = TRUE, nu = 50, p.val = 0.01)
}
\arguments{
\item{expr}{a data matrix with cells in the columns and genes in the rows, preferably standardized gene-wise}

\item{sample}{TRUE/FALSE, if the parameters should be estimated by random sampling or not, default is FALSE}

\item{cor}{TRUE/FALSE, if the svd should be calculated on the correlation matrix or not (covariance matrix), default is TRUE}

\item{nu}{the number of gene singular vectors to calculate in the process (the more, the more time expansive), default is 50}

\item{p.val}{the p-value to be used in the test of normality for the singular vectors, default is 0.01}
}
\value{
A MP object:
\itemize{
\item eigen: eigenvalues and vectors of the cell-cell correlation matrix
\item maxEigen: maximum eigenvalue of the MP distribution
\item minEigen: minimum eigenvalue of the MP distribution
\item sig_vectors: singular vectors that lie significantly above the MP distribution
\item M: the number of genes
\item N: the number of cells
\item svd: the singular value decomposition
\item genes.used: genes that have been used for the calculation of the MP distribution
\item p.value_mp_fit: p-value for the similarity of a MP distribution to the eigenvalues found as noise
\item transcriptome_mode: the index of the transcriptome mode
\item input_parameters: the inputs to the function
}
}
\description{
This function takes a data matrix as input and out puts the parameters asscociated to the MP distribution.
}
\examples{
library(splatter)
data("splatO")
expr <- counts(splatO)
expr <- expr[rowSums(expr)>0,]

#Normalize and log-transform the data
expr.norm <- t(t(expr)/colSums(expr))*10000
expr.norm.log <- log(expr.norm + 1)

expr.scale <- t(scale(t(expr.norm.log)))
L <- fit_mp(expr.scale)

}
