% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get informaton about each singular vectos}
\usage{
get_info(measure_obj, cluster)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}

\item{cluster}{the name of the cluster of interest, e.g. "Pod"}
}
\value{
A data.frame with the following information:
\itemize{
\item phiclust: the value of sigma for each singular value
\item g_phiclust: the value of g-sigma for each sigular value
\item theta: the singular values of the signal matrix
\item r2vals: the adjusted r squared obtained during the regression
\item singular_value: the index of the singular values (from highest to lowest after the first calculation)
\item celltype: the selected cluster
}
}
\description{
This function extracts information about sigma in a cluster.
}
\examples{
data("out")
get_info(out, "Group2")

}
