% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_genes.R
\name{get_var_genes}
\alias{get_var_genes}
\title{Extract variance driving genes}
\usage{
get_var_genes(measure_obj, cluster)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}

\item{cluster}{the name of the cluster of interest, e.g. "Pod"}
}
\value{
A matrix, with the columns as the number of significant gene singular vecotrs and the rows contain the 1\% genes with highest and lowest values in the gene singular vector
}
\description{
This function lists the variance driving genes for each significant singular vector
}
\examples{
data("out")
get_var_genes(out, "Group2")

}
