% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\docType{data}
\name{out}
\alias{out}
\title{Measure object created by applying sigma_funct}
\format{
A measure object
}
\usage{
out
}
\description{
This object contains all elements as described in the output of sigma_funct
}
\keyword{datasets}
