% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kidney.R
\docType{data}
\name{paga.coord}
\alias{paga.coord}
\title{Force-directed graph layout}
\format{
data.frame
}
\source{
produced with scanpy
}
\usage{
paga.coord
}
\description{
This force-drected graph layout has been generated with scanpy in python. For detailed description of the method to genrate this layout, have a look at the methods part of our paper.
}
\keyword{datasets}
