% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_g_phiclust.R
\name{plot_all_g_phiclusts}
\alias{plot_all_g_phiclusts}
\title{Plot all g_phiclusts}
\usage{
plot_all_g_phiclusts(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}
}
\value{
ggplot plot
}
\description{
plot of all g-phiclusts obtained per cluster
}
\examples{
data("out")
plot_all_g_phiclusts(out)

}
