% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_sigmas.R
\name{plot_all_phiclusts}
\alias{plot_all_phiclusts}
\title{Plot all phiclusts}
\usage{
plot_all_phiclusts(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}
}
\value{
ggplot plot
}
\description{
plot of all phiclusts obtained per cluster
}
\examples{
data("out")
plot_all_phiclusts(out)

}
