% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mp.R
\name{plot_MP}
\alias{plot_MP}
\title{Plot fit of MP}
\usage{
plot_MP(measure_obj, cluster, plot.title = "")
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}

\item{cluster}{the name of the cluster of interest, e.g. "Pod"}

\item{plot.title}{optional title of the plot}
}
\value{
ggplot plot
}
\description{
singular values density plot and MP density as a solid line
}
\examples{
data("out")
plot_MP(out, "Group2")

}
