% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phiclust.R
\name{plot_phiclust}
\alias{plot_phiclust}
\title{Plot phiclust}
\usage{
plot_phiclust(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}
}
\value{
ggplot plot
}
\description{
plot the value of phiclust for every cluster
}
\examples{
data("out")
plot_phiclust(out)

}
