% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singular_vectors.R
\name{plot_singular_vectors}
\alias{plot_singular_vectors}
\title{Plot singular vectors for a cluster}
\usage{
plot_singular_vectors(
  measure_obj,
  cluster,
  v1 = 1,
  v2 = 2,
  colour = NULL,
  scaled = FALSE
)
}
\arguments{
\item{measure_obj}{the measure-object as the output of phiclust}

\item{cluster}{the name of the cluster of interest, e.g. "Pod"}

\item{v1}{index of vector to plot on the x-axis, default is 1}

\item{v2}{index of vector to plot on the y-axis, default is 1}

\item{colour}{either a gene name or a vector the same length as the number of cells, default is NULL (no colour)}

\item{scaled}{TRUE/FALSE, if the scaled expression of a gene should be shown, default is FALSE}
}
\value{
ggplot plot
}
\description{
scatter plot of two singular vectors for a selected cluster
}
\examples{
data("out")
plot_singular_vectors(out, "Group2")

}
