% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kidney.R
\docType{data}
\name{rb.genes}
\alias{rb.genes}
\title{List of ribosomal genes}
\format{
character string
}
\source{
HGNC website
}
\usage{
rb.genes
}
\description{
We used this list of ribosomal genes to exclude these variances from the calculation of the measure.
}
\keyword{datasets}
