% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kidney.R
\docType{data}
\name{sce_kidney}
\alias{sce_kidney}
\title{SingleCellObject of W16 fetal kidney}
\format{
Single Cell Object
}
\source{
\url{https://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.3000152}
}
\usage{
sce_kidney
}
\description{
This data set contains a SingleCellobject of a kidney of week 16 of gestation. The authors identified 22 clusters in total.
}
\keyword{datasets}
