% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\docType{data}
\name{splatO}
\alias{splatO}
\title{SingleCellObject generated from splatter}
\format{
Single Cell Object
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-017-1305-0}
}
\usage{
splatO
}
\description{
This object contains a simulation od scRNAseq data with 5 clusters of with the following parameters:
nGenes = 1000, batchCells = 500, group.prob = c(0.1, 0.2, 0.3, 0.1, 0.3), path.nonlinearProb = 0, out.prob = 0.1,
dropout.type = "none", lib.loc = 11, de.downProb = 0.5, de.prob = 0.05, seed = 3, de.facLoc = 0.1
}
\keyword{datasets}
