% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kidney.R
\docType{data}
\name{stress.genes}
\alias{stress.genes}
\title{List of stress genes}
\format{
string of characters
}
\source{
\url{https://www.nature.com/articles/nmeth.4437?proof=t}
}
\usage{
stress.genes
}
\description{
We used this list of stress genes to exclude these variances from the calculation of the measure in the human fetal kidney data set.
}
\keyword{datasets}
